refactor: 상수 파일 세분화
