+++ src/constants/path.js
@@ -1,0 +1,8 @@
+const PATH = {
+  ROOT: '/',
+  PRODUCTS: 'products',
+  CARTS: 'carts',
+  ORDERS: 'orders',
+};
+
+export default PATH;
